rrr: convert rpc/channel to module (rrr.channel)
channel.cpp is the rrr.channel module interface unit; .hpp deleted.
Pure interface header: ChannelError enum, ChannelFrame POD,
ChannelConnectionBase/ChannelListenerBase/ChannelFactoryBase virtual
interfaces, On*Callback typedefs (via detail::CallbackWrapper from
rrr.callback_wrapper), and ConnectResult/proxy typedefs.

Includers (6) updated:
  - rpc/{fiber_channel,tcp_channel,inmemory_channel}.hpp,
    rpc/{client,server}.hpp, rrr.hpp.

One gotcha: client.hpp uses `detail::CallbackWrapper<>` directly
(line 323 — FutureAttr's reply_callback_ field) but previously got it
transitively through `channel.hpp -> callback_wrapper.hpp`. After
the conversion, `import rrr.channel;` does not re-export
callback_wrapper, so client.hpp needed an explicit
`import rrr.callback_wrapper;` added.

This is the "module imports are not transitive by default" gotcha
worth documenting: switching to `import` from `#include` breaks
transitive-include chains that consumers may have implicitly relied
on. Each consumer must import what it directly uses.

27 modules. Tree green.
