+++ src/rrr/CMakeLists.txt
@@ -32,6 +32,7 @@ set(RRR_MODULE_SRC
     ${CMAKE_CURRENT_SOURCE_DIR}/misc/stat.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/reactor/epoll_wrapper.cc
     ${CMAKE_CURRENT_SOURCE_DIR}/rpc/callbacks.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/rpc/channel.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/rpc/circuit_breaker.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/rpc/connection_metrics.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/rpc/connection_state.cpp
