+++ src/rrr/rpc/client.hpp
@@ -30,11 +30,12 @@
 import rrr.epoll_wrapper;
 
 
-#include "channel.hpp"
 #include "fiber_channel.hpp"
 #include "request_queue.hpp"
 
+import rrr.callback_wrapper;
 import rrr.callbacks;
+import rrr.channel;
 import rrr.circuit_breaker;
 import rrr.connection_metrics;
 import rrr.connection_state;
